# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateResourceRequest:
    properties:
      capacity:
        minimum: 1
        type: integer
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        enum:
        - MEETING_ROOM
        - CAR
        - DEVICE
        type: string
    required:
    - capacity
    - name
    - type
    type: object
...
  dto.ResourceResponse:
    properties:
      capacity:
        type: integer
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
...
  dto.UpdateResourceRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  handler.CapacityInput:
    properties:
      delta:
        minimum: 1
        type: integer
    required:
    - delta
    type: object
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.IDResponse:
    properties:
      id:
        type: string
    type: object
  handler.StatusResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for authentication, resources, and bookings.
  title: Diploma API
  version: "1.0"
paths:
...
  /resources/:
    get:
      description: Returns all resources.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ResourceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Creates a new resource. Requires ADMIN role.
      parameters:
      - description: Resource payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create resource
      tags:
      - resources
  /resources/{id}:
    delete:
      description: Deletes resource by UUID. Requires ADMIN role.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete resource
      tags:
      - resources
    get:
      description: Returns one resource by UUID.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get resource by ID
      tags:
      - resources
    put:
      consumes:
      - application/json
      description: Updates resource fields. Requires ADMIN role.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Resource update payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update resource
      tags:
      - resources
  /resources/{id}/capacity/decrease:
    patch:
      consumes:
      - application/json
      description: Decreases resource capacity by delta. Requires ADMIN role.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Capacity delta
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CapacityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Decrease resource capacity
      tags:
      - resources
  /resources/{id}/capacity/increase:
    patch:
      consumes:
      - application/json
      description: Increases resource capacity by delta. Requires ADMIN role.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Capacity delta
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CapacityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Increase resource capacity
      tags:
      - resources
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"